Factor AbstractSubmissionIT out of SubmissionProcessorIT. * Creates a future EmptySubmissionIT to be implemented * Updates SubmissionProcessorIT to inherit from AbstractSubmissionIT * Common test methods went in to SubmissionTestUtil * Import cleanup
